+++ src/main/cpp/query_table/query_table.h
@@ -70,7 +70,7 @@ namespace  vaultdb {
             unsigned int getTupleCount() const;
 
             std::string toString(const bool &showDummies = false) const;
-            std::string toString(const int & limit, const bool &showDummies = false) const;
+            std::string toString(const size_t & limit, const bool &showDummies = false) const;
 
             void putTuple(const int &idx, const QueryTuple<B> &tuple);
